backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
asgiref==3.10.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
autopep8==2.3.2
bandit==1.8.6
black==25.9.0
bqplot==0.12.45
branca==0.8.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
colour==0.1.5
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
Django==5.2.7
django-cachalot==2.8.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.3
django-geojson==4.1.0
django-redis==6.0.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
earthengine-api==0.1.358
eerepr==0.1.2
exceptiongroup==1.3.0
executing==2.2.1
Fiona==1.9.4
flake8==7.3.0
folium==0.20.0
fonttools==4.60.0
frozenlist==1.8.0
future==1.0.0
geemap==0.24.0
geocoder==1.38.1
geojson==3.0.1
geopandas==0.14.0
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==2.16.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
httplib2==0.31.0
idna==3.10
inflection==0.5.1
ipyevents==2.0.4
ipyfilechooser==0.6.0
ipyleaflet==0.20.0
ipython==8.37.0
ipytree==0.2.2
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-leaflet==0.20.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.7.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==1.24.3
openai==0.28.1
packaging==25.0
pandas==2.1.0
parso==0.8.5
pathspec==0.12.1
Pillow==10.0.1
platformdirs==4.4.0
plotly==5.17.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.7
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.0.9
pyperclip==1.10.0
pyproj==3.7.1
pyshp==2.3.1
python-box==7.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.2
ratelim==0.1.6
redis==6.4.0
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.3.0
scipy==1.11.2
scooby==0.10.1
seaborn==0.13.0
shapely==2.0.1
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
stevedore==5.5.0
tenacity==9.1.2
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
tracker==0.1.1
traitlets==5.14.3
traittypes==0.2.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
xyzservices==2025.4.0
yarl==1.22.0
